services/database_management/requirements.txt
google-api-python-client
gdoctableapppy
Flask
gunicorn
Flask_Cors
pandas
requests
aiohttp
python-dotenv